// dragonball-app/src/app/interfaces/planets.interface.ts
export interface PlanetsData {
    items: planet[];
    meta:  Meta;
    links: Links;
}

export interface planet {
    id:          number;
    name:        string;
    isDestroyed: boolean;
    description: string;
    image:       string;
    deletedAt:   null;
}

export interface Links {
    first:    string;
    previous: string;
    next:     string;
    last:     string;
}

export interface Meta {
    totalItems:   number;
    itemCount:    number;
    itemsPerPage: number;
    totalPages:   number;
    currentPage:  number;
}